refactor/fix dash collision in panels
